chore(core): rename `aws-sdk` dir to `aws_sdk`
