+++ packages/core/src/httpAuthSchemes/aws_sdk/AWSSDKSigV4Signer.ts
