+++ packages/core/src/httpAuthSchemes/aws_sdk/index.ts
