+++ packages/core/src/httpAuthSchemes/aws_sdk/resolveAWSSDKSigV4Config.ts
