+++ packages/core/src/httpAuthSchemes/aws_sdk/throwAWSSDKSigningPropertyError.ts
