+++ packages/core/src/httpAuthSchemes/index.ts
@@ -1,1 +1,1 @@
-export * from "./aws-sdk";
+export * from "./aws_sdk";
